+++ Bonobo.Git.Server/Data/Repository.cs
@@ -1,9 +1,21 @@
+using Bonobo.Git.Server.App_GlobalResources;
 using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Linq;
 
 namespace Bonobo.Git.Server.Data
 {
+    public enum RepositoryPushMode
+    {
+        [Display(ResourceType = typeof(Resources), Name = "No")]
+        No = 0,
+        [Display(ResourceType = typeof(Resources), Name = "Yes")]
+        Yes,
+        [Display(ResourceType = typeof(Resources), Name = "Global")]
+        Global,
+    }
+
     public partial class Repository
     {
         private ICollection<Team> _teams;
@@ -16,6 +28,7 @@ namespace Bonobo.Git.Server.Data
         public string Description { get; set; }
         public bool Anonymous { get; set; }
         public byte[] Logo { get; set; }
+        public RepositoryPushMode AllowAnonymousPush { get; set; }
 
         public virtual ICollection<Team> Teams
         {
@@ -54,8 +67,7 @@ namespace Bonobo.Git.Server.Data
         }
 
         public bool AuditPushUser { get; set; }
-
-
+        
         /// <summary>
         /// Correct a repository name have the same case as it has in the database
         /// If the repo is not in the database, then the name is returned unchanged