+++ Bonobo.Git.Server/Models/RepositoryModels.cs
@@ -25,12 +25,18 @@ namespace Bonobo.Git.Server.Models
         public string Group { get; set; }
         public string Description { get; set; }
         public bool AnonymousAccess { get; set; }
+        public RepositoryPushMode AllowAnonymousPush { get; set; }
         public UserModel[] Users { get; set; }
         public UserModel[] Administrators { get; set; }
         public TeamModel[] Teams { get; set; }
         public bool AuditPushUser { get; set; }
         public byte[] Logo { get; set; }
         public bool RemoveLogo { get; set; }
+
+        public RepositoryModel()
+        {
+            AllowAnonymousPush = RepositoryPushMode.Global;
+        }
 
         public bool NameIsValid
         {
@@ -73,6 +79,11 @@ namespace Bonobo.Git.Server.Models
 
     public class RepositoryDetailModel
     {
+        public RepositoryDetailModel()
+        {
+            AllowAnonymousPush = RepositoryPushMode.Global;
+        }
+
         public Guid Id { get; set; }
 
         [UniqueRepoName]
@@ -113,6 +124,9 @@ namespace Bonobo.Git.Server.Models
 
         [Display(ResourceType = typeof(Resources), Name = "Repository_Detail_Anonymous")]
         public bool AllowAnonymous { get; set; }
+
+        [Display(ResourceType = typeof(Resources), Name = "Repository_Detail_AllowAnonymousPush")]
+        public RepositoryPushMode AllowAnonymousPush { get; set; }
 
         [Display(ResourceType = typeof(Resources), Name = "Repository_Detail_Status")]
         public RepositoryDetailStatus Status { get; set; }